Create README.md
When an R package developer wishes to publish their package on CRAN, they must undergo a validation process. This process involves a series of tests conducted on the package to verify that it complies with CRAN standards. One of these tests is cross-validation, which involves running the R CMD check command on different platforms. The time it takes to execute this command is recorded in the check_times.csv file, which contains data from over 13,000 packages, including the respective execution time on each platform and the characteristics of each package, such as the number of dependencies, the number of files in the src directory, the number of files in the data directory, etc.

The objective of this project is to analyze the different characteristics of R packages and determine which of these characteristics influence the execution time of the R CMD check command. Additionally, it aims to determine if it is possible to predict the execution time of this command using the package's characteristics.